tr/http.py: determine the correct ping URL automatically.
The trick here is to try connecting to the ACS.  After doing
socket.connect(), our local socket will have done an implicit bind() to the
right local IP (if there's more than one) for the interface used to connect
to the ACS.  Since we use that interface to connect to the ACS, then the ACS
should use that interface to connect back to us, which means the IP address
of that interface is the right one for the ping URL.  We can get it by just
going getsockname() on the outgoing socket (regardless of whether it
successfully connects, ie. this still works even if the ACS is down right
now).