*/fix_path.py: centralize sys.path fixups.
There was a lot of ad-hoc adding of stuff to sys.path in a bunch of files,
and it didn't really work reliably when the additions were relative to $PWD
but a program might be run as /path/to/program from a different directory.

Clean it up by putting a fix_path.py in each directory, and having all
python files in each directory import it.  They all eventually include
tr.fix_path that includes the right vendor/* directories.