cwmpd: force line buffering of stdout/stderr.
When redirecting to something other than a tty, python was "helpfully"
switching to full buffering instead of line buffering, so messages wouldn't
show up right away.  This occurred when doing things like

	cwmpd 2>&1 | logger -t cwmpd

...which is annoying.  We force line buffering so it flushes the buffer
after each line, which is what happens when stdout is a tty.

(Probably we should stop using 'print' and start calling a logging function,
which would auto-flush after every message.  But this is okay for now.)