Adding tr157.
Adds SetParameterAttributes into api_soap.py and api.py.
Fixes DeleteObject.
Add a PeriodicStatistics object onto InternetGatewayDevice.

Work is in progress on collecting statistics, but so
far you can add objects, and configure the parameters
and at least on statistic is collected.  If notification
is enabled, a new session will be started with the
event '4 VALUE CHANGED'.

Still TODO:
- Save the sample times on both the SampleSet and the Parameter
- Handle the time objects and not just SampleInterval.
- Handle the various ways to collect data:
  o Average
  o Max
  o Min
  o Value (currently only current value is done)

Change-Id: Ia19065a91b985b7df83ff51c714d5be1f1e1d273