gfmedia: clean up redundant class name prefix/suffixes.
Java programmers get used to naming things this.that.ThisThatThing, because
when you import it, it'll be named "ThisThatThing" in your local namespace,
dropping the this.that. prefix.  So if you just named it "Thing" you might
run into conflicts.

In python this isn't a problem, so it's fine (and much more readable) to
just call it this.that.Thing.

This is especially valuable in implementing cwmp data models because this
way you can name the struct after the cwmp model it's actually implementing,
so you can have platform.gfmedia.device.Device,
platform.fakecpe.device.Device, etc.

This patch cleans up a few of those in platform/gfmedia.  There are others
remaining in various other directories, but it's a start.

Change-Id: Ifb0fac46a94bc8b8afac3a188ebcaaa5826400c3