download/installer: make it more resilient, and add a cli command.
If certain things went wrong with a download, it would leave a tmp file in
/tmp, because we disabled NamedTemporaryFile's auto-delete feature.
Re-enable that feature (and keep the NamedTemporaryFile object around to
make sure it doesn't delete it until we want it to).

Don't redirect ginstall's output to /dev/null, because that's just
intensely cruel to anyone trying to debug problems.

Add a new "download" command to the cwmp command-line interface, so we can
test downloading features.

Change-Id: Ibaf5ef4a990c7b3767642fe7151ae16bf741fca4