download: handle cases where config_dir isn't writable.
This isn't great, because we can't save and restore the download state.  But
it should be enough to allow a download+install to complete and the system
to reboot, even if we can't notify the ACS that it was successful
afterwards.

It's non-optimal, but in cases where bugs in the system image prevent the
config_dir from working at all, upgrading, even non-optimal upgrading, may
be the only way to fix it.

While we're here, redirect the download_dir from /rw/tr69/dnld to /tmp/dnld.
/tmp may be more likely to fill up than /rw, but at least a simple reboot
will fix /tmp, but flash might stay full and prevent future upgrades.

Change-Id: Ia0e2251405de3a2bf91a9d4f68b9427b520047ff