Monitor fan speed
sysmgr now controls fan speed more aggressively, and will stop the
fan altogether if the system temperature is low enough.

Its actually somewhat difficult to get a fan started again. Setting
the fan to its minimum voltage isn't enough to get it moving, you have
to set it to a higher voltage first and then lower it down to the speed
you want.

Provide a way to monitor fan speed, just in case. By proxy we will
also be monitoring the noise produced within the home, which is good
to have data for.

Change-Id: Icb6450cd5d77294717b152f6dfb680ea66c2a0ab