Fix race condition in ProcessStatus
There are more race conditions in ProcessStatus than previously realized.
For example if we use GetParameterNames, we get a list of the processes
which exist at that moment. Returning a SOAP:Fault for processes in that
list which no longer exist isn't great.

Make up a new process status for exited processes.

Also correct the datatypes. PID, Priority, and Size are all supposed
to be integers not strings.

Change-Id: Ie3ffb171964c4c2e25990721a73a3819bceca606