Compatibility with monotonic-time-capable tornado.
The next patch changes tornado to work with time.monotonic().  This patch
changes catawampus to be compatible with both the new and the old tornado,
while we wait for (hopefully) our patch to be accepted upstream.

Mostly this is done by using timedelta() where we should have been using it
anyway.  But there are a few places (particularly unit tests) where we have
to take two different code branches.

Change-Id: I7958c84a1c5b94725afbf56202db9099266e5dec