Renumber tr-135 object instances
We had been incrementing the object instance for each new MPEG stream.
When first powered on, the first MPEG stream would be:
Device.Services.STBService.1.ServiceMonitoring.MainStream.1.*

After changing channels 98 times, the MPEG stream would be:
Device.Services.STBService.1.ServiceMonitoring.MainStream.99.*

This made the STBService impossible to use with PeriodicStatistics.

Instead, number the object instances from 1 to N (where N is the maximum number
of streams). Export the stream_id as a X_GOOGLE-COM_StreamID parameter instead,
to allow analysis of long-term video statistics.

Instance number are persistent through the life of the stream. For example if
we're monitoring 8 streams, they will be numbered 1..8. If streams 1 through 7
go away, there will be one stream left and it will be numbered 8. If two new
streams appear the set of MainStream instances will be [1, 2, 8]

Change-Id: Ibf7e5efb6154e382d8e7196e22eb571f3d3d13af