tr/http.py: limit maximum message dump length.
We historically have logged all sent and received tr-069 xml messages to
stdout.  This is super useful for debugging, but as time has passed, these
messages are getting more and more full of stuff.  Let's trim out the middle
of long messages to we get the gist, but don't overflow the log buffer.

Change-Id: Iffede035473c1d5849db7a36685468cb45d0f0bf