tr/types: add @x.validator decorators.
Instead of overriding setters, we introduce the concept of a "validator",
which is a tentative setter: you give it a value, and it raises an exception
if the attribute wouldn't be settable to that value, or returns the value (or
a slightly adjusted or type-coerced value) if the set would succeed.  Then
we implement the setter as an operation that calls the validator and, if
there's no exception, sets it to the returned value.

You can't 100% trust validators: not every attribute has them,
and something that's valid one moment may no longer be valid the next.  So
transactions still need to be able to support rollbacks, but with validators
we should be able to avoid them *most* of the time at least.

As a bonus, validators are actually less coding than getters + setters.
Usually your getter is trivial, but you have to write one just so you can
write a setter.  Validators let you write just the validator, leaving
getters and setters implicit.

Change-Id: I5e1a1f5f73861ffdd9d51f6d33c2798743f97cda