tr/core: add LookupExports() and reduce unnecessary object lookups.
It's time to remove the unnecessary calls to ValidateExports() and write
actual tests to make sure we don't traverse the tree any more than needed.

LookupExports() is like FindExport(), except it can look up a bunch of names
at once, and doesn't traverse the entire tree for each entry, so it should
be much faster (and moreover, if you have dynamically-generated objects,
it will only need to generate the object once to show multiple parameter
values).

Change-Id: Ife84d0cd7d983ac4120ee59a621248d11a7604ce