Export extended TCPStats.
Existing TCPStats implementation was done speculatively:
there were no producers of the JSON files for TCPStats.
Now that there are, adjust the implementation and unit
tests to match what they produce and extend the implementation
to add the new fields they include.

Also: finally give up and treat a Python long as an xsd:unsignedInt.
This isn't strictly correct, as a SOAP unsignedInt is supposed
to be limited to 32 bit values, but we're now exporting a large
byte count which has no other reasonable encoding. We'll expect that
ACSes we talk to will handle 64 bit numbers.

Change-Id: I10c3c686e0c71c4174b514764a14ed281d533a99