Wifi device for FakeCPE, with tests.
Reasonably accurate simulation of a Wifi device. It enforces valid
enumerated values for setting its parameters, and simulates some aspects
of the behavior of the real device (like disabling AutoChanEnable when
a channel is explicitly set).

It makes no attempt to simulate the operation of the actual Wifi network.
AssociatedDevices don't come and go, none of the stats increment, etc.

Change-Id: Ie4ef143f2d38b4d6cecb797e5788762c0aa30840