Improve GFiberTV data structure.
This maps the SageTV properties to "real" TR-069 properties in a tree that
looks more like a normal TR-069 configuration structure than the old Mailbox
arrangement.  The main limitation of the Mailbox structure is the ACS needs
to know how to treat it specially if it wants to set more than one value,
since setting multiple values requires toggling the Name field, then the
Value field, then the Name field again, etc.

While we're here, use all the modern catawampus amenities, and clean up a
few of them:

- Remove the manual transaction management code: the built-in catawampus
  core transactions work fine for this use case.  This makes the code much
  simpler and less error prone.

- Create a new tr.types.FileBacked() type that automatically gets and sets
  values from the contents of a file instead of storing them in memory.

- Use tr.types for all parameters instead of writing our own getters and
  setters.

- Add a new tr.types.Date that parses TR-069 or time_t style dates and
  produces datetime.datetime objects.  Teach api_soap.Soapify() how to
  render datetime objects back into TR-069 dates.

- Bonus: teach Soapify() how to render signed vs. unsigned integers as
  different xsd:type strings, even though I'm not using that for anything
  in this patch.

Change-Id: Ief50e973f6ee97dde49c22dc2b7f5561f1f1f0e5