GFiberTv: don't crash if /app/sage/Sage.properties.defaults is missing.
This would happen if SageTV hasn't started yet, because that file "exists"
but is a symlink to another file in /tmp, which hasn't been created.  So the
glob.glob() passes but the file still can't be opened.  This sounds like it
would be a rare race condition, but unfortunately it happens every time if
NTP hasn't synced yet.

We were caching the list of properties forever, which when combined with
this fix, could cause a race condition that would prevent the config from
ever being populated at all.  So change it to reread the file once per ACS
session instead.  That way once the file gets created, all is well.

Change-Id: I0072eef37e60331d27be03f3017cda8fb3f99b5a