Add a nonstandard AddObjects TR-069 command.
This is just like AddObject, but it creates multiple objects in one call.
This can greatly reduce the number of TR-069 turnarounds if you want to add
multiple objects.

The command looks like this:
	<cwmp:AddObjects>
	  <ObjectName>Fun.Object.Name.</ObjectName>
	  <Count>3</Count>
	  <ParameterKey>MyParameterKey567</ParameterKey>
	</cwmp:AddObjects>

and the response looks like this:
	<cwmp:AddObjectsResponse>
	  <InstanceNumber>123</InstanceNumber>
	  <InstanceNumber>124</InstanceNumber>
	  <InstanceNumber>125</InstanceNumber>
	  <Status>0</Status>
	</cwmp:AddObjectsResponse>

You can check for the existence of this command by using the (standard)
GetRPCMethods command, or just try to run it and get an error message back
if it doesn't exist.

Since this command is non-standard, I considered using an X- prefix in its
name, but that seems to be discouraged practice nowadays,
eg: http://tools.ietf.org/html/rfc6648#appendix-B .  In short, if you use
the X- prefix and then standardize it without an X- prefix later, you end up
having to support both forever, so there's no benefit.

Change-Id: I124298acc8ec754c98947a12fd9dcf0791f1c0a2