Change AddObjects command to X_CATAWAMPUS_ORG_AddObjects.
The new name is considerably uglier and doesn't follow the advice of the
IETF about naming stuff.  On the other hand, it complies with the TR
standard for extensions.  And after all, it's TR-069, so a "little" uglier
won't make a measurable difference really.

We also add the ability to AddObjects to more than one object at a time.
The new command looks like this:

The command looks like this:
	<cwmp:X_CATAWAMPUS_ORG_AddObjects>
	  <Object>
	    <ObjectName>Fun.Object.Name.</ObjectName>
	    <Count>3</Count>
	  </Object>
	  <Object>
	    <ObjectName>Fun.Object2.</ObjectName>
	    <Count>1</Count>
	  </Object>
	  <ParameterKey>MyParameterKey567</ParameterKey>
	</cwmp:X_CATAWAMPUS_ORG_AddObjects>

and the response looks like this:
	<cwmp:X_CATAWAMPUS_ORG_AddObjectsResponse>
	  <Object>
            <ObjectName>Fun.Object.Name.</ObjectName>
            <InstanceNumber>123</InstanceNumber>
            <InstanceNumber>124</InstanceNumber>
            <InstanceNumber>125</InstanceNumber>
          </Object>
          <Object>
            <ObjectName>Fun.Object2.</ObjectName>
            <InstanceNumber>42</InstanceNumber>
          </Object>
	  <Status>0</Status>
	</cwmp:X_CATAWAMPUS_ORG_AddObjectsResponse>

If any errors are encountered, the successfully-added objects are deleted
and an X_CATAWAMPUS_ORG_AddObjectsFault is generated, in the same style as
SetParameterValuesFault.  You can see a sample of how to examine the results
in acs_integration_test.py.

Change-Id: I0904d527768289e2105dd0cf0d85cdc00db9bf6d