Add i2c to catawampus/tr/vendor.
In order to enable import of i2c module by dm/ds6923_optical.py.
i2c is still being installed, so it can be used by sfmodule.
Had to modify fix_path.py to add path.
This is done so that catawampus can be used as a standalone open
source project.

Change-Id: Icfe83bc7bb660e76d9962f052378364a8df1a222