Fix support for HTTP 301 and HTTP 302 responses from ACS.
It's common practice in http, when you receive an 301 or 302 response to a
POST, to redirect to the new location with a GET instead.  This is to
prevent accidentally POSTing twice (since POST requests are intended to do
work, you wouldn't want to accidentally double-submit a form).  This
protection isn't useful in TR-069 however, since TR-069 is meaningless
without POST.  So we set the pycurl option to enable POST upon 301 or 302
redirect.  libcurl always does POST on 307, so that doesn't need special
treatment.

Testing it was another matter.  It turned out http_test.py wasn't actually
using a real http client, so we couldn't correctly test the behaviour.  Upon
enabling that, I found a bug (apparently a real one) in libcurl that caused
the redirect test to fail.  You need libcurl 7.31 or above, which fixes the
actual bug.  (It was closing sockets *before* unregistering from epoll,
which creates a race condition.) In case you don't have that version, we
work around the bug (although it really just hides the problem and it's not
a good choice for production use) by disabling epoll.  Because this isn't a
real fix, we print a warning on startup if it's needed.

Change-Id: If9215352ec7b76b74c317462d7b5faa15c738771