Clean up some old-style and nonstandard sh syntax.
It hasn't been necessary to use the preceding 'x' in sh string comparisons
for well over a decade.  Let's remove all the examples of it in order not to
encourage anyone else.

Also change "if [...] ; then" to "if [...]; then".  This is less of a
problem, but the latter is more common, so let's keep it consistent.

Change-Id: I1f7d5bc9ad5d8c74f746f735f6158437c2672648