periodic_statistics: remember calls to GetCanonicalName on SampleSets.
GetCanonicalName is very inefficient and ends up instantiating (up to) the
entire object tree in order to find values.  And the canonical name of an
object never changes (objects are created and destroyed, but not
moved around), so it's safe to cache it forever.

We also print a message upon handling each SampleSet, in case calculating
the SampleSet results in an error, so at least we'll know what we were
handling at the time.

Change-Id: Ib34c71ba63c3989b607bfe1e1e955f32c2293a77