brcmwifi: increase autochannel delay so it works with 5 GHz.
Fixes b/10950954.  'wl autochannel' scans only the 2.4 GHz or 5 GHz
channels, depending which channel it's already on.  That is, if you do
'wl channel 1' and then autochannel, it'll scan 2.4 GHz channels.  If you do
'wl channel 52' and then autochannel, it'll scan 5 GHz channels.

There are more 5 GHz channels so, unsurprisingly, it takes longer to scan
them all.  In my test, with repeated tries, it takes about 1000ms to scan
2.4 GHz channels, and 2500ms to scan the 5 GHz channels.  This makes sense
since there are 11 2.4 GHz channels and 24 5 GHz channels if you run 'wl
channels'.  So a 3-second sleep should be enough.

Change-Id: Ieeeb41bb523e5108dfe9bc0fb51a4834318d6fa9