/config/nickname, yo.
ACS populates DeviceProperties on TV Boxes now. We don't really
need to know the nicknames of other devices, but store the NickName
of this device in /config/nickname. It is stored in flash (not /tmp)
because Nicknames basically never change once set, and dialserver
needs it to be set at startup.

We'll use it in dialserver, so the TV Box will show up in the list
using its nickname.

We'll probably also use it in dhclient, to set the DHCP HostName.

Fixes b/10151485

Change-Id: I1fce86794503c82edd21113f63a3e7efcca5f428