Add tr.types.NumberOf
diagui expects to add notifiers to Device.Ethernet.InterfaceNumerOfEntries.
For that to work, it has to descend from tr.types.Attr. Now I suspect that
monitoring InterfaceNumerOfEntries is just arbitrary, and that in reality
we're not checking if a new Ethernet interface suddenly materialized in
the system, but I wanted to add a NumberOf type anyway to handle a common
bit of boilerplate. So here we are.

Change-Id: I2f20c9d307b8e224405465a715fb3760cd7eaba6