brcmwifi: fix auto channel selection, sleeps, and OperatingFrequencyBand.
WL_SLEEP was not accomplishing anything useful.  It works equally well with
and without the sleep.  The only necessary sleep is when waiting for
autochannel to complete.

OperatingFrequencyBand and Channel could get into conflict, where a channel
was selected that was intended for the other band.  Now, if you set a
channel, set OperatingFrequencyBand to the right thing for that channel.  If
you set OperatingFrequencyBand and Channel isn't in the right range, set
AutoChannelEnable to make the configuration valid.  (If you subsequently set
the Channel to something valid, that's fine; setting Channel always disables
AutoChannelEnable.)

While we're here, change the way the autochannel implementation works.  We
will probably never use it because the autochannel implementation in the
driver is a bit hopeless (there is no way to disable DFS channels for
example) but it might as well work.  The problem was that 'autochannel 2'
doesn't actually work reliably, and you almost always end up (silently of
course) using a different channel than autochannel would have selected.
Which could explain a thing or two.  Instead, read the channel it selected,
and set it explicitly at the right time in the AP startup sequence.

Change-Id: Iddd6079fb9cee126fd845d28811e8bd745e944c5