Make the index of interfaces reliably reproducible for host.py.
The interfaces are being indexed by order, but os.listdir doesn't seem
to order by name. Sort them so that we know the order is deterministic.

This makes the test pass on my machine. It was already passing on
others. In theory, it also makes us less vulnerable to ordering in general.

Change-Id: I4d53c0dc4b3c6b8374af2112ecce297b012a7c47