Fix rounding in CalcTimeToNextSample
Inside this braintwister of a function we calculate the number of
seconds to wait before taking a new sample.  Sometimes we round off and
collect before the next interval, and end up with sample times like:
4,1,4,1,4,1  where we'd expect 5,5,5,5,5

The adding in of a 1/2 here should kick the round to the next integer in
those cases so we'll see the expected sequence.

Change-Id: I1aaf062bcee8fba980009c07345d89ddd5857ac3