WAN interfaces are Device.IP.Interface.4 and 5.
We've renamed interfaces several times (for example,
'ath0' and 'ath1' became 'wlan0' and 'wlan1'), and this
resulted in changing the object numbering of Device.IP.Interface.

ACS needs stable numbering. Fix the order of the IP interfaces
to what the ACS expects. Note this means there can be holes
in the numbering, a platform might have object indexes 1, 2, 4, 5.
Gaps are fine for the ACS, renumbering is the problem.

Change-Id: Ic3ba3b5874815191a37d8c9e61cff11e6c4ca728