Convert to using moca2json.
Parsing the output of mocap consumes quite a bit of CPU
time. Profiling shows that Device.MoCA is by far the most
expensive set of periodic statistics we collect, in terms
of CPU time spent parsing and time spent waiting for the
cwmpd<->mocap<->mocad<->firmware round trips.

Additionally, Broadcom has several times changed the output
of mocap in ways which subtly broke the catawampus data model.
mocap is not intended to be scraped, it is for human use.

Instead, rely on a moca2json process to periodically query
for MoCA status and write its results to JSON files. These
are much less expensive to parse, and do not block waiting
for a response from the MoCA device. Additionally as moca2json
uses the MoCA 2.x SDK, we shouldn't have subtle breakage
from output format changes. Significant changes should
cause moca2json compilation to fail, so we'll know about
it and can fix it.

Change-Id: Ibae9671e8cd31efb0a99e94a8d3492cbfc457977