Simplify and fix tr.types.NumberOf().
The way it was written had a few problems:

- self.listobj was shared between all instances of a given class.  So
  if you created two instances of the same datamodel object containing
  a WhateverList, both {a,b}.WhateverNumberOfEntries would reflect
  len(b.WhateverList).

- It had to be initialized with an awkward
  type(self).WhateverListNumberOfEntries.SetList() call.

- If you reassigned self.WhateverList, WhateverListNumberOfEntries would
  still retain a reference to the *old* WhateverList, and give the wrong
  number of entries.

Also, simplify the implementation by inheriting from ReadOnly() instead of
Attr(), so we don't have to re-implement the read-only feature.

Add a check to ValidateExports() to confirm that xNumberOfEntries matches
len(x), so prevent this kind of accident in the future.

Change-Id: Id6201303d2a60fa3a959a46b8c31596d1068ea31