dm/device_info: add caching to ProcessStatus object.
An earlier fix to validate that len(xNumberOfEntries) matches len(x) was
triggering for this case, because the number of processes on the system kept
changing.  In general, caching the process list for a while is probably the
right thing to do anyway, or becomes hard to manipulate the data.

This uncovered a problem with tr.session.cache where it can't cache
generator objects, but would silently fail in weird ways.  Catch that
problem explicitly, and add an @tr.session.cache_as_list to make it easy to
work around.

Change-Id: Ia99335d44f6a0a088d85b0572fa2767116645056