cwmp client: optimize parsing of quotedblock objects.
This was responsible for about 50% of the time taken to retrieve a fairly
large "cwmp rlist /" on our device.  In my case, it takes it from around 90
seconds down to around 40 seconds, with 9 of those seconds still taken up by
client-side parsing (versus 59 or so before).

This could be faster still if we replaced the protocol with something more
binary, or maybe json, both of which have python native ways to do the
parsing.

Note: this code only affects the cwmp command-line client protocol, which is
not really used for anything in production but which we want to use in
automated tests.  So it should be safe to backport to a branch if we want.

Change-Id: Icfe3eb3048f7236d56176943bee35c93fe47e57a