fd #17, you don't exist. Go away.
If we get an event for an fd which no longer has a handler,
there is no useful action we can take. Rather than sit in a
tight loop raising KeyError exceptions as fast as we can,
just close the fd and unregister it from the eventfd.

Reduces severity of b/12896999 to tolerable levels.

Change-Id: I1c4890e55c1029b46dbe6b7420fc4f64384c9940