Pass Standard to /bin/wifi
/bin/wifi defaults to allowing ac, but some of the other
options supported in ac don't work on ath9k (like A-MPDU
EXP3). We want ath9k to function as an 802.11n device, so
tell it to do so.

Also allow the Standard to be set from the ACS, and not be read-only.
With GFRG1x0 we found that some houses got better coverage by turning
off 'n' and leaving only 'g'. Making Standard writable will allow
the same thing to be done for GFRG2x0, if necessary.

Fixes b/17317261.

Change-Id: Ifc9ad8a619ded84c5fcdde7c3a7ff4415532049d