I have no idea what I'm doing...
Somewhere in the bowels of tornado, something wants the clock to advance
or it never does a poll() on the socket that returned EAGAIN.  I
honestly have no idea why this fixes the test on gTrusty or why the test
works on gPrecise.

I stumbled on this because enable tracing with:
python -m trace
Adds enough delays that the tests pass so I figured it was something
with timing, but since the test passes with tracing enabled it's sort of
difficult to divine where it's getting hung up with tracing disabled.

Change-Id: Ic43dd9f3d1a1f516a9fe10abc78e9deea8bcf076