Split parts of tr.core.Exporter into a new tr.handle.Handle class.
The idea here is to remove the more complicated bits of functionality (tree
navigation) from Exporter objects in the device tree, into a "Handle" object
that "holds onto" the device tree and lets you manipulate it.  Most files
that implement the device model don't need any of the Handle functionality,
so this helps keep things separate, and discourage using the Handle layer
from inside the device model (instead of doing the more obvious thing: just
accessing the member variables).

This is a cleanup I've been wanting to do for quite a while.  It's
especially important now so that we can have a clean separation of the
Experiment functionality (which needs to introduce a slightly fictional
layer between the "real" device model and what is reported to the server).

Change-Id: I33bf90bc8d6191a41199734117c371dac455467b