Trivial beginning of experiment framework.
This implements a very simple data model that will allow you to register
experiments, which modify various parameters when applied.  You can then
enable/disable them using TR-069 parameters.

This version is trivial (in that it doesn't do anything with the variables).
But it does correctly implement the API, so code running this version will
appear to the ACS as if it supports the experiment API, but has no
experiments registered.

Change-Id: Ib55efa9d17b8dc0aa519d9edee5582e7764b370f