Rearrange 'make lint' rule to run more than one py file per invocation.
The previous workaround takes much longer to execute because of how slowly
gpylint starts up.  Since it runs as part of every 'make test', that really
slows down the debug cycle.

The root cause of the problem was that gpylint would become confused if you
lint a file with the same name from two different directories in the same
invocation.  To avoid this problem, break it up per directory.

Example minimal test steps that confuse gpylint, which this patch avoids:
    gpylint --disable=g-bad-import-order --disable=g-unknown-interpreter \
        platform/tomato/device.py \
        platform/gfmedia/device_test.py

Change-Id: Ibf9fed2e93ad21a0808ab5ee631629eca0b2480a