parse-schema: Support path= parameter for <component> tags.
Some of the schemas import components at the toplevel but address them to
objects that are *under* the toplevel using a path= attribute that we were
ignoring.  Let's just ignore the question of why they couldn't just import
them using a <component> object inside the object they were actually
affecting.

By fixing the parser, we answer the mystery in a TODO of why the UPnP object
was not being auto-exported like it should be.

Unfortunately, fixing this also auto-exports a whole bunch more objects that
should have been exported all along but were previously being dumped into
the wrong classes and thus ignored, so we have to now unexport all of those.

After that's done, we correct a few places where an extension schema was
defined incorrectly as if it were "only" a list of extensions; it's actually
an inheritence tree, so if you're using the extension object, you shouldn't
also instantiate the base class.  x-cata181.xml was the main victim here.
To avoid confusion, update it to tr181 v2.6, since other places in the code
were expecting v2.6.

Change-Id: Ife892a376c19db2e65a0c7a363d836d28db5051b