Collect catawampus extension objects under Device.X_CATAWAMPUS-ORG.
We were starting to accrue inconveniently-named extension objects in a few
different places in the tree.  According to the tr-069 set of standards, we
should not actually have multiple objects at the root level.  So let's get
started cleaning this up.  This patch moves all the extension objects to a
single tr-181 compliant location under Device.X_CATAWAMPUS-ORG.

For backward compatibility, we still leave aliases in place at the original
object locations.  Once all devices in the field are upgraded, we can drop
the old locations.

We also move all the toplevel extension objects into dm/ instead of
platform/gfmedia.  This allows them to be tested as part of platform/fakecpe.

Change-Id: Ice4588723a72cba4e636a3b5c4009e7ba74d3fa6