Disable super-init-not-called lint warning.
Unfortunately this warning triggers randomly depending on the collection of
files that happen to be included in the lint command line.  I put a simple
test case into b/18440404.  If gpylint gets fixed, we can re-enable this
warning (which would otherwise be very useful, since forgetting to call a
superclass init is pretty common).

While we're here, change the makefile to sort files alphabetically before
sending them to xargs.  This really shouldn't be needed, but if gpylint is
going to be flakey depending on file order, and file order depends on your
local filesystem, let's do this to minimize lint discrepancies between
developers and buildbots.

Change-Id: If052f25350eb0d723f060c77a4d211a6619dc47c