tr.cwmptypes.FileBacked: avoid create/delete mixup condition.
@tr.mainloop.WaitUntilIdle will remember each *variant* of a given function
call, including the parameters passed.  So if you tried to change the value
of a delete_if_empty FileBacked attribute to a nonempty string, and then an
empty string, or vice versa, it would schedule a rename *and* a delete, in
some order.  What you actually wanted was exactly one thing to happen,
corresponding to the most recent change to the attribute.

To fix it, remove all parameters from _ReallyWriteFile() and just have it
use the existence or nonexistence of its input temp file as the clue for
whether its output should exist or not.

While we're here, fix some erroneous superclass constructor calls in
_Proxy and FileBacked.  There was no way to actually make these cause a
problem (and thus I can't make a test for it), but they were definitely wrong.

Change-Id: Idf1710642f0f817894cbec4241107a019277b168