cwmpd: enable FileBacked notifiers and add FileBacked.SetFileName()
These passed in tests, but weren't enabled in the actual cwmpd yet.  Turn
them on, and fix some bugs that ensued.

SetFileName() lets you override a FileBacked attribute on a per-instance
basis.  Until now, all FileBacked attributes have effectively been
singletons.  We're going to introduce some that include the parent object
name as part of the path, so we have to be able to change it per object.

Change-Id: I6396572448c531fa4622e87fc466a1c4ebda29e0