tr/core and tr/handle: support alternative Exporter implementations.
The default core.Exporter works if you know in advance what parameters your
object has, which is by far the most common case.  But if you want to
instantiate them on the fly when people get/set them, we might need a bit
more flexibility.

Break out an AbstractExporter base class to allow for alternative
implementations.

Change-Id: Ibb31e7951a5c2370aa49e6aab6f1cb890768e3bf