experiment.py: add support for "system-level" experiments.
System-level experiments are implemented outside cwmpd, somewhere inside the
operating system's boot scripts.  They are expected to persist across
restarts of catawampus (possibly across reboots).  Catawampus can only
request that such an experiment be enabled or disabled, but has to wait for
the OS itself to finish the enabling or disabling process, which might
happen only after rebooting.

Thus, for such an experiment it's possible for it to (for a while) be both
Available and Requested but not Active, or Available and Active but not
Requested.

Change-Id: I521aca073f649bdd5cc88c523e73aa3ae2556f03