dnsmasq: one line per IP
If there are multiple MAC addresses and/or ClientIDs
which are all supposed to map to one IP address, they
must all go on one line.

In our environment in particular, this is a reasonable thing
to want to do. We bridge Wifi and Ethernet onto a single L2
topology, its natural to want the laptop to get the same
IP address whether its using the Wifi interface or Ethernet.

Fixes b/18906872

While here, suppress a log message from dnsmasq on every ACS
checkin. If the ACS URL did not change, we don't do anything
so don't log anything.

Change-Id: I03117d4fd6456dfd2aeeb09ca371c3a1ed73f75e