Add a schema for mapping an external address to an internal address.
The support for SMB spec says that we will map five external public IPs to
five different internal IPs.  Using the existing NAT interface is
insufficient since that maps an external interface and not an
external IP on the external interface.

Also using this model is not needed to populate the external interface
with the public IP.

Change-Id: I3906348547e3f827b32fdd2c33ca4aa344bcf4d4