wifiblaster: add wifiblaster data model.
This adds a data model to interface with wifiblaster with four
variables: Duration, Enable, Interval, and Size. Duration and Size
control the wifiblaster duration and packet size, respectively. Enable
enables or disables wifiblaster. Interval controls the average time
between packet blasts. The variables are communicated to waveguide over
the filesystem.

Change-Id: Ie8b5a7a8ba1bb9d31c25fb096630e1c1f32a9578