Add data model for iBeacons.
iBeacon is a Bluetooth Low Energy service which lets
a mobile device react when an iBeacon comes into range.
The beacon includes a measured RSSI at 1 meter, allowing
the device to check the RSSI it sees from the beacon to
determine how far away it must therefore be. So the mobile
app can figure out which TV Box it is physically closest to.

Change-Id: I20d8a29076b6575bcb12a6f8b663a757da0d5ed3