Allow more frequent pings
Way, way back we thought the ACS would ping the device once and
then keep the CWMP session open for as long as it needed to.
That didn't turn out to be feasible, the way the ACS backend
instances work. The ACS needs to let the CWMP session end and
then ping it again in the possibly near future.

Change-Id: I3718523f7fefff7ec9dd47c0134315514247e5dc