Fix some new warnings with the new gpylint version.
Note: one side effect of using .keys() in iteration is that it generates a
temporary list, so that if the dict changes during iteration, the keys()
list is not affected.  This is sometimes what you want, and sometimes not.

For the change in api.py, it looks like the dict should never change during
iteration, so it's safe to remove .keys().  For tr/session.py, the iteration
deliberately mutates the dict, so we can't just drop .keys(), but it turns
out we can just .clear() instead and skip the iteration.

Change-Id: I6dae725aca5eb831e4d628f3be8daf77c980c846